Add files via upload
PileOfLemons is the file that holds everything together and is the one that should be ran. It will gather all the files off of smogon.com and then call the other classes to do the rest.

ParseLeads parses through the simple leads and usage files. making them 'pretty'
ParseMoveset is a big more complicated but parses through the moveset file. making them 'pretty'
CombinePretty takes the pretty files and combines them into one file. 

Parsestats does the ParseLeads and ParseMoveset and Combinepretty. But accepts a date range and goes month by month generating this files.

ReportsPoke makes a simple bar graph based off the info in the combined pretty files.


This is all very simple and can be modified easily. The combined.json files make everything even easier.